repair type of block_size

+++ examples/l2fwd-crypto/main.c
@@ -179,7 +179,8 @@ struct l2fwd_crypto_params {
 
 //	unsigned digest_length; as the rte_pktmbuf_append need uint16_t
 	uint16_t digest_length;
-	unsigned block_size;
+//	unsigned block_size;
+	uint16_t block_size;
 
 	struct l2fwd_key iv;
 	struct l2fwd_key aad;
